리뷰 별점 출력

+++ src/main/java/com/spharosacademy/project/SSGBack/cart/dto/Output/CartOutputDto.java
@@ -21,7 +21,7 @@ public class CartOutputDto {
         private String productBrand;
         private Long count;
         private float price;
-        private float qty;
+        private int qty;
         private String titleImgUrl;
 
         OptionCartOutputDto optionCartOutputDto;
